+++ refinedstorage-autocrafting-api/src/main/java/com/refinedmods/refinedstorage/api/autocrafting/calculation/NumberOverflowDuringCalculationException.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.api.autocrafting.calculation;
 
-public class NumberOverflowDuringCalculationException extends RuntimeException {
+public class NumberOverflowDuringCalculationException extends CalculationException {
     NumberOverflowDuringCalculationException() {
         super("Invalid amount");
     }