+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/preview/AutocraftingPreviewContainerMenu.java
@@ -2,10 +2,12 @@ package com.refinedmods.refinedstorage.common.autocrafting.preview;
 
 import com.refinedmods.refinedstorage.api.autocrafting.preview.Preview;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
+import com.refinedmods.refinedstorage.common.api.support.resource.PlatformResourceKey;
 import com.refinedmods.refinedstorage.common.api.support.resource.ResourceContainer;
 import com.refinedmods.refinedstorage.common.support.containermenu.AbstractResourceContainerMenu;
 import com.refinedmods.refinedstorage.common.support.containermenu.DisabledResourceSlot;
 import com.refinedmods.refinedstorage.common.support.containermenu.ResourceSlotType;
+import com.refinedmods.refinedstorage.common.support.packet.c2s.C2SPackets;
 import com.refinedmods.refinedstorage.common.support.resource.ResourceContainerImpl;
 
 import java.util.ArrayList;
@@ -102,6 +104,21 @@ public class AutocraftingPreviewContainerMenu extends AbstractResourceContainerM
         }
         if (!last) {
             setCurrentRequest(requests.getFirst());
+        }
+    }
+
+    public void maxAmountResponseReceived(final long maxAmount) {
+        if (listener == null) {
+            return;
+        }
+        if (currentRequest.getResource() instanceof PlatformResourceKey resource) {
+            listener.maxAmountReceived(resource.getResourceType().getDisplayAmount(maxAmount));
+        }
+    }
+
+    void requestMaxAmount() {
+        if (currentRequest.getResource() instanceof PlatformResourceKey resource) {
+            C2SPackets.sendAutocraftingPreviewMaxAmountRequest(resource);
         }
     }
 }